" Rebase latest updates with master branch
Build of this repo with maven has been observed
to fail when using JDK 11, it however passes with
JDK 8.

This was discovered to be due to the way the maven
compiler processes the Jenkins 'Extension' annotation
between the different JDK versions.

More details can be found here at the maven compiler
Jira page: https://bit.ly/2OIzHf1
"

"Rebase local repo with master branch

Squashed all previous local commits into one
and rebased them on the project's master
branch to resolve earlier conflicts."

Add RequestContext class

Add UserCreationListener extensionpoint class

Update UserCreationListener extensionpoint class

Change file indentation of java files

Change HeaderPrefix context to reflect jenkins

Change java file indentation

Change java  file indentation

Change file indentation of java file

Update java file indentation

Change file indentation of java file

Add Apache Log4j dependencies to pom.xml

Update Apache Log4j dependencies in pom.xml

Update Apache Log4j dependencies in pom.xml

Verify successful maven validation of plugin

Verify successful compilation upto 'mvn test-compile'

Refactor project parent pom.xml

Refactor project parent pom.xml

Refactor project parent pom.xml

Refactor project parent pom.xml

Add log4j-audit RequestContext class

Update log4j-audit RequestContext class

Update pom.xml

"Remove requestId attribute from login-event

The login event attributes that are of particular interest
to be logged are the 'timestamp' and 'userId'.
In contrast, the 'requestId' is a general attribute information
that is common throughout the application and it is logged via
the RequestContext methods."

"Remove requestId attribute from logout-event

The logout event attributes that are of particular interest
to be logged are the 'timestamp' and 'userId'.
In contrast, the 'requestId' is a general attribute information
that is common throughout the application and it is logged via
the RequestContext methods."

"Remove request attributes from createUser event

The createUser event attributes that are of particular interest
to be logged are the name of the new user, 'userId' and the
time the creation was made, 'timestamp'.
In contrast, the 'requestId' and 'requestUri' are general
attribute information that are common throughout the application
and they can be logged via the RequestContext methods."

"Remove request attributes from updateKey event

The updateKey event attributes that are of particular interest
to be logged are the name of the user whose key was updated,
'userId' and the time the update was made, 'timestamp'.
In contrast, the 'requestId' and 'requestUri' are general
attribute information that are common throughout the application
and they can be logged via the RequestContext methods."

"Remove updatePassword event request attributes

The updatePassword event attributes that are of particular interest
to be logged are the name of the user whose password was updated,
'userId' and the time the password update was made, 'timestamp'.
In contrast, the 'requestId' and 'requestUri' are general
attribute information that are common throughout the application
and they can be logged via the RequestContext methods."

"Update timestamp event attribute

Argueably the timestamp attribute can be said to be a general
event attribute throughout the application, which can be logged
via the RequestContext methods, however the attribute is still
of particular interest to time-sensitive audit events (such as
'login', 'logout', and 'updatePassword') that require explicit
logging of the time of their occurence."

Add UserLogListener

Refactor UserLogListener class

Update NODE_NAME to 'master'

Refactor RequestContext.java class

Refactor UserLogListener.java class

Update UserLogListener.java class